Extract SongBuilder
Extract building a song from (parsed) data and keeping track of state to a builder.
