+++ src/chord_sheet/song.ts
@@ -1,4 +1,4 @@
-import Line, { LineType } from './line';
+import Line from './line';
 import Paragraph from './paragraph';
 import Key from '../key';
 import ChordLyricsPair from './chord_lyrics_pair';
@@ -6,51 +6,16 @@ import Metadata from './metadata';
 import ParserWarning from '../parser/parser_warning';
 import MetadataAccessors from './metadata_accessors';
 import Item from './item';
-import TraceInfo from './trace_info';
-import FontStack from './font_stack';
-
-import {
-  ABC,
-  BRIDGE, CHORUS, GRID, LILYPOND, Modifier, NONE, ParagraphType, TAB, VERSE,
-} from '../constants';
+import { CHORUS, Modifier } from '../constants';
 
 import Tag, {
   CAPO,
-  CHORUS as CHORUS_TAG, END_OF_ABC,
-  END_OF_BRIDGE,
+  CHORUS as CHORUS_TAG,
   END_OF_CHORUS,
-  END_OF_GRID, END_OF_LY,
-  END_OF_TAB,
-  END_OF_VERSE,
   KEY,
-  NEW_KEY, START_OF_ABC,
-  START_OF_BRIDGE,
   START_OF_CHORUS,
-  START_OF_GRID, START_OF_LY,
-  START_OF_TAB,
-  START_OF_VERSE,
-  TRANSPOSE,
 } from './tag';
-
-const START_TAG_TO_SECTION_TYPE = {
-  [START_OF_ABC]: ABC,
-  [START_OF_BRIDGE]: BRIDGE,
-  [START_OF_CHORUS]: CHORUS,
-  [START_OF_GRID]: GRID,
-  [START_OF_LY]: LILYPOND,
-  [START_OF_TAB]: TAB,
-  [START_OF_VERSE]: VERSE,
-};
-
-const END_TAG_TO_SECTION_TYPE = {
-  [END_OF_ABC]: ABC,
-  [END_OF_BRIDGE]: BRIDGE,
-  [END_OF_CHORUS]: CHORUS,
-  [END_OF_GRID]: GRID,
-  [END_OF_LY]: LILYPOND,
-  [END_OF_TAB]: TAB,
-  [END_OF_VERSE]: VERSE,
-};
+import SongBuilder from '../song_builder';
 
 type MapItemsCallback = (_item: Item) => Item | null;
 
@@ -77,16 +42,6 @@ class Song extends MetadataAccessors {
 
   _bodyParagraphs: Paragraph[] | null = null;
 
-  currentKey: string | null = null;
-
-  currentLine: Line | null = null;
-
-  fontStack: FontStack = new FontStack();
-
-  sectionType: ParagraphType = NONE;
-
-  transposeKey: string | null = null;
-
   warnings: ParserWarning[] = [];
 
   /**
@@ -96,16 +51,6 @@ class Song extends MetadataAccessors {
   constructor(metadata = {}) {
     super();
     this.metadata = new Metadata(metadata);
-  }
-
-  get previousLine(): Line | null {
-    const count = this.lines.length;
-
-    if (count >= 2) {
-      return this.lines[count - 2];
-    }
-
-    return null;
   }
 
   /**
@@ -143,32 +88,6 @@ class Song extends MetadataAccessors {
     }
 
     return copy;
-  }
-
-  chords(chr: string): void {
-    if (!this.currentLine) throw new Error('Expected this.currentLine to be present');
-    this.currentLine.chords(chr);
-  }
-
-  lyrics(chr: string): void {
-    this.ensureLine();
-    if (!this.currentLine) throw new Error('Expected this.currentLine to be present');
-    this.currentLine.lyrics(chr);
-  }
-
-  addLine(line?: Line): Line {
-    if (line) {
-      this.currentLine = line;
-    } else {
-      this.currentLine = new Line();
-      this.lines.push(this.currentLine);
-    }
-
-    this.setCurrentProperties(this.sectionType);
-    this.currentLine.transposeKey = this.transposeKey ?? this.currentKey;
-    this.currentLine.key = this.currentKey || this.metadata.getSingle(KEY);
-    this.currentLine.lineNumber = this.lines.length - 1;
-    return this.currentLine;
   }
 
   private expandLine(line: Line): Line[] {
@@ -255,90 +174,6 @@ class Song extends MetadataAccessors {
     });
 
     return paragraphs;
-  }
-
-  setCurrentProperties(sectionType: ParagraphType): void {
-    if (!this.currentLine) throw new Error('Expected this.currentLine to be present');
-
-    this.currentLine.type = sectionType as LineType;
-    this.currentLine.textFont = this.fontStack.textFont.clone();
-    this.currentLine.chordFont = this.fontStack.chordFont.clone();
-  }
-
-  ensureLine(): void {
-    if (this.currentLine === null) {
-      this.addLine();
-    }
-  }
-
-  addTag(tagContents: string | Tag): Tag {
-    const tag = Tag.parseOrFail(tagContents);
-    this.applyTagOnSong(tag);
-    this.applyTagOnLine(tag);
-    return tag;
-  }
-
-  private applyTagOnLine(tag: Tag) {
-    this.ensureLine();
-    if (!this.currentLine) throw new Error('Expected this.currentLine to be present');
-    this.currentLine.addTag(tag);
-  }
-
-  private applyTagOnSong(tag: Tag) {
-    if (tag.isMetaTag()) {
-      this.setMetadata(tag.name, tag.value || '');
-    } else if (tag.name === TRANSPOSE) {
-      this.transposeKey = tag.value;
-    } else if (tag.name === NEW_KEY) {
-      this.currentKey = tag.value;
-    } else if (tag.isSectionDelimiter()) {
-      this.setSectionTypeFromTag(tag);
-    } else if (tag.isInlineFontTag()) {
-      this.fontStack.applyTag(tag);
-    }
-  }
-
-  setSectionTypeFromTag(tag: Tag): void {
-    if (tag.name in START_TAG_TO_SECTION_TYPE) {
-      this.startSection(START_TAG_TO_SECTION_TYPE[tag.name], tag);
-      return;
-    }
-
-    if (tag.name in END_TAG_TO_SECTION_TYPE) {
-      this.endSection(END_TAG_TO_SECTION_TYPE[tag.name], tag);
-    }
-  }
-
-  startSection(sectionType: ParagraphType, tag: Tag): void {
-    this.checkCurrentSectionType(NONE, tag);
-    this.sectionType = sectionType;
-    this.setCurrentProperties(sectionType);
-  }
-
-  endSection(sectionType: ParagraphType, tag: Tag): void {
-    this.checkCurrentSectionType(sectionType, tag);
-    this.sectionType = NONE;
-  }
-
-  checkCurrentSectionType(sectionType: ParagraphType, tag: Tag): void {
-    if (this.sectionType !== sectionType) {
-      this.addWarning(`Unexpected tag {${tag.originalName}}, current section is: ${this.sectionType}`, tag);
-    }
-  }
-
-  addWarning(message: string, { line, column }: TraceInfo): void {
-    const warning = new ParserWarning(message, line || null, column || null);
-    this.warnings.push(warning);
-  }
-
-  addItem(item: Item): void {
-    if (item instanceof Tag) {
-      this.addTag(item);
-    } else {
-      this.ensureLine();
-      if (!this.currentLine) throw new Error('Expected this.currentLine to be present');
-      this.currentLine.addItem(item);
-    }
   }
 
   /**
@@ -539,6 +374,10 @@ Or set the song key before changing key:
     return clonedSong;
   }
 
+  addLine(line: Line) {
+    this.lines.push(line);
+  }
+
   /**
    * Change the song contents inline. Return a new {@link Item} to replace it. Return `null` to remove it.
    * @example
@@ -555,15 +394,16 @@ Or set the song key before changing key:
    */
   mapItems(func: MapItemsCallback): Song {
     const clonedSong = new Song();
+    const builder = new SongBuilder(clonedSong);
 
     this.lines.forEach((line) => {
-      clonedSong.addLine();
+      builder.addLine();
 
       line.items.forEach((item) => {
         const changedItem = func(item);
 
         if (changedItem) {
-          clonedSong.addItem(changedItem);
+          builder.addItem(changedItem);
         }
       });
     });
@@ -587,13 +427,14 @@ Or set the song key before changing key:
    */
   mapLines(func: MapLinesCallback): Song {
     const clonedSong = new Song();
+    const builder = new SongBuilder(clonedSong);
 
     this.lines.forEach((line) => {
       const changedLine = func(line);
 
       if (changedLine) {
-        clonedSong.addLine();
-        changedLine.items.forEach((item) => clonedSong.addItem(item));
+        builder.addLine();
+        changedLine.items.forEach((item) => builder.addItem(item));
       }
     });
