+++ src/parser/chord_sheet_parser.ts
@@ -3,6 +3,7 @@ import Song from '../chord_sheet/song';
 import Line from '../chord_sheet/line';
 import ChordLyricsPair from '../chord_sheet/chord_lyrics_pair';
 import { deprecate, normalizeLineEndings } from '../utilities';
+import SongBuilder from '../song_builder';
 
 const WHITE_SPACE = /\s/;
 const CHORD_LINE_REGEX = /^\s*((([A-G|Do|Re|Mi|Fa|Sol|La|Si])(#|b)?([^/\s]*)(\/([A-G|Do|Re|Mi|Fa|Sol|La|Si])(#|b)?)?)(\s|$)+)+(\s|$)+/;
@@ -22,6 +23,8 @@ class ChordSheetParser {
   preserveWhitespace = true;
 
   song: Song = new Song();
+
+  songBuilder: SongBuilder = new SongBuilder(this.song);
 
   songLine: Line | null = null;
 
@@ -79,7 +82,7 @@ class ChordSheetParser {
   endOfSong() { }
 
   parseLine(line) {
-    this.songLine = this.song.addLine();
+    this.songLine = this.songBuilder.addLine();
 
     if (line.trim().length === 0) {
       this.chordLyricsPair = null;
@@ -135,7 +138,8 @@ class ChordSheetParser {
     }
 
     if (!lyricsLine.trim().length) {
-      this.songLine = this.song.addLine();
+      this.songLine = new Line();
+      this.song.addLine(this.songLine);
     }
   }
